Split tmvar into v1 and v2
- Two different folders (tmvar_v1, tmvar_v2)
- Changed citation entry for tmvar_v1
